feat(repo): add @uipath/flow-node-schema package
Extract Zod schemas from @uipath/apollo-react/canvas into a standalone
@uipath/flow-node-schema package with only zod as a dependency. This
allows CLI tools and non-React consumers to validate flow/canvas
definitions without pulling in React and all UI dependencies.
