+++ packages/apollo-react/src/canvas/schema/node-instance/index.ts
@@ -1,13 +1,19 @@
-// Re-export types
 export type {
   InstanceDisplayConfig,
   InstanceId,
   InstanceTypeVersionKey,
   InstanceVersion,
-} from './base';
-// Re-export schemas
-export { displayConfigSchema, idSchema, typeVersionKeySchema, versionSchema } from './base';
-export type { EdgeInstance } from './edge';
-export { edgeSchema } from './edge';
-export type { InstanceUiConfig, NodeInstance } from './node';
-export { nodeSchema, uiSchema } from './node';
+  EdgeInstance,
+  InstanceUiConfig,
+  NodeInstance,
+} from '@uipath/flow-node-schema';
+
+export {
+  displayConfigSchema,
+  idSchema,
+  typeVersionKeySchema,
+  versionSchema,
+  edgeSchema,
+  nodeSchema,
+  uiSchema,
+} from '@uipath/flow-node-schema';
