+++ packages/flow-node-schema/rslib.config.ts
@@ -1,0 +1,39 @@
+import { defineConfig } from '@rslib/core';
+
+export default defineConfig({
+  lib: [
+    {
+      format: 'esm',
+      output: {
+        distPath: {
+          root: './dist',
+        },
+        filename: {
+          js: '[name].js',
+        },
+      },
+      dts: true,
+    },
+    {
+      format: 'cjs',
+      output: {
+        distPath: {
+          root: './dist',
+        },
+        filename: {
+          js: '[name].cjs',
+        },
+      },
+      dts: false,
+    },
+  ],
+  source: {
+    entry: {
+      index: './src/index.ts',
+    },
+  },
+  output: {
+    target: 'web',
+    cleanDistPath: true,
+  },
+});
