+++ packages/flow-node-schema/src/index.ts
@@ -1,0 +1,3 @@
+export * from './node-definition';
+export * from './node-instance';
+export * from './toolbar';
