+++ packages/flow-node-schema/src/node-definition/index.ts
@@ -1,0 +1,31 @@
+export type { CategoryManifest } from './category-manifest';
+export { categoryManifestSchema } from './category-manifest';
+export type { ConnectionConstraint, HandleTarget } from './constraints';
+export { connectionConstraintSchema, handleTargetSchema } from './constraints';
+// Re-export types
+export type {
+  HandleCategory,
+  HandleConfigurationSpecificPosition,
+  HandleGroupManifest,
+  HandleGroupOverride,
+  HandleManifest,
+  HandlePosition,
+  HandleType,
+} from './handle';
+// Re-export schemas
+export {
+  handleConfigurationSpecificPositionSchema,
+  handleGroupManifestSchema,
+  handleGroupOverrideSchema,
+  handleManifestSchema,
+  handlePositionSchema,
+  handleTypeDisplaySchema,
+  handleTypeSchema,
+} from './handle';
+export type { FormSchemaLike, NodeDisplayManifest, NodeManifest, NodeShape } from './node-manifest';
+export {
+  nodeDebugManifestSchema,
+  nodeDisplayManifestSchema,
+  nodeManifestSchema,
+  nodeShapeSchema,
+} from './node-manifest';
