+++ packages/flow-node-schema/src/node-instance/edge.ts
@@ -1,0 +1,25 @@
+import { z } from 'zod';
+import { idSchema } from './base';
+
+/**
+ * A connection between nodes in a workflow
+ */
+export const edgeSchema = z.object({
+  id: idSchema,
+  sourceNodeId: idSchema,
+  /** The source port name (output port) */
+  sourcePort: z.string().min(1),
+  targetNodeId: idSchema,
+  /** The target port name (input port) */
+  targetPort: z.string().min(1),
+  data: z
+    .object({
+      /** Optional label displayed at the edge midpoint */
+      label: z.string().optional(),
+    })
+    .passthrough()
+    .optional(),
+});
+
+// Export inferred TypeScript type
+export type EdgeInstance = z.infer<typeof edgeSchema>;
