+++ packages/flow-node-schema/src/node-instance/index.ts
@@ -1,0 +1,13 @@
+// Re-export types
+export type {
+  InstanceDisplayConfig,
+  InstanceId,
+  InstanceTypeVersionKey,
+  InstanceVersion,
+} from './base';
+// Re-export schemas
+export { displayConfigSchema, idSchema, typeVersionKeySchema, versionSchema } from './base';
+export type { EdgeInstance } from './edge';
+export { edgeSchema } from './edge';
+export type { InstanceUiConfig, NodeInstance } from './node';
+export { nodeSchema, uiSchema } from './node';
